feat: make DbClient's message table month optional
(for BigTable)

also:
- kill save_channels (only needed by legacy autoconnect)
- update mockall/cleanup old workaround
- update openssl per RUSTSEC-2023-0044

SYNC-3449

+++ autoconnect/autoconnect-ws/autoconnect-ws-sm/src/identified/mod.rs
@@ -256,32 +256,33 @@ pub async fn process_existing_user(
     app_state: &Arc<AppState>,
     user: &User,
 ) -> DbResult<Option<ClientFlags>> {
-    // TODO: current_month likely not applicable on BigTable. autoendpoint also
-    // has this code
-    let Some(ref current_month) = user.current_month else {
-        debug!("Missing `current_month` value, dropping user"; "user" => ?user);
-        app_state
-            .metrics
-            .incr_with_tags("ua.expiration")
-            .with_tag("errno", "104")
-            .send();
-        app_state.db.remove_user(&user.uaid).await?;
-        return Ok(None);
-    };
+    // TODO: if BigTable, can we assume the user is migrated at this point (so
+    // we wouldn't need to validate `current_month` is not None?)
+    if let Some(rotating_message_table) = app_state.db.rotating_message_table() {
+        let Some(ref current_month) = user.current_month else {
+            debug!("Missing `current_month` value, dropping user"; "user" => ?user);
+            app_state
+                .metrics
+                .incr_with_tags("ua.expiration")
+                .with_tag("errno", "104")
+                .send();
+            app_state.db.remove_user(&user.uaid).await?;
+            return Ok(None);
+        };
 
-    let message_table = app_state.db.message_table();
-    if current_month != message_table {
-        debug!("User is inactive, dropping user";
-            "db.message_table" => message_table,
-            "user.current_month" => current_month,
-            "user" => ?user);
-        app_state
-            .metrics
-            .incr_with_tags("ua.expiration")
-            .with_tag("errno", "105")
-            .send();
-        app_state.db.remove_user(&user.uaid).await?;
-        return Ok(None);
+        if current_month != rotating_message_table {
+            debug!("User is inactive, dropping user";
+                   "db.message_table" => rotating_message_table,
+                   "user.current_month" => current_month,
+                   "user" => ?user);
+            app_state
+                .metrics
+                .incr_with_tags("ua.expiration")
+                .with_tag("errno", "105")
+                .send();
+            app_state.db.remove_user(&user.uaid).await?;
+            return Ok(None);
+        }
     }
 
     let flags = ClientFlags {
